requirements.txt
pandas==1.1.0
statsmodels==0.11.1
scikit-learn==0.23.2
mysql-connector==2.2.9
paho-mqtt==1.5.0